tg_bot/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
catboost==1.2.2
certifi==2023.11.17
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
frozenlist==1.4.1
graphviz==0.20.1
idna==3.6
imageio==2.33.1
kiwisolver==1.4.5
lazy_loader==0.3
magic-filter==1.0.12
matplotlib==3.8.2
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
plotly==5.18.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
tenacity==8.2.3
tifffile==2023.12.9
typing_extensions==4.9.0
tzdata==2023.4
yarl==1.9.4
